requirements.txt
django_paypal==0.5.0
social_auth_core==1.7.0
python_social_auth==0.3.6
Django==2.0.5
django_jsonfield==1.0.1
paypal==1.2.5
social==0.1.dev3
